Fold health-gate specs into the HarmonizeCommand spec file
